Partial move of units into its own directory.
